Create Issue templates for Upgrade regressisons (#105)
* Create Issue templates for Upgrade regressisons, Pick Requests and Test Reports.

These will serve as issue templates for our GitHub Project based Release management. Once finalized, we will link to these from https://github.com/facebook/react-native/issues/new/choose

The Upgrade Regression form is a 1:1 copy of the form in the facebook/react-native repository, and seeks to replace that, directing the issues to this repository instead.

The Pick Request Form aims to replace the freeform comments requesting picks in eg https://github.com/reactwg/react-native-releases/discussions/104, so that we can make use of the Project Management flows to track the progress and outcome of these requests.

The Test Report Form aims to replace the test progress reporting in `#releases-crew` on Discord. This is currently a limited visibility destination, making test reports public and structure in issues may help triaging and root causing issues later on.

The issues generated via these forms will be ingested to the appropriate projects via Workflows. Those are to follow ASAP.
To protect this repository against spam issues in other catagories, the `config.yml` disallows freeform issue creation in this repo.

* Create Issue templates for Upgrade regressisons, Pick Requests and Test Reports.

These will serve as issue templates for our GitHub Project based Release management. Once finalized, we will link to these from https://github.com/facebook/react-native/issues/new/choose

The Upgrade Regression form is a 1:1 copy of the form in the facebook/react-native repository, and seeks to replace that, directing the issues to this repository instead.

The Pick Request Form aims to replace the freeform comments requesting picks in eg https://github.com/reactwg/react-native-releases/discussions/104, so that we can make use of the Project Management flows to track the progress and outcome of these requests.

The Test Report Form aims to replace the test progress reporting in `#releases-crew` on Discord. This is currently a limited visibility destination, making test reports public and structure in issues may help triaging and root causing issues later on.

The issues generated via these forms will be ingested to the appropriate projects via Workflows. Those are to follow ASAP.
To protect this repository against spam issues in other catagories, the `config.yml` disallows freeform issue creation in this repo.